+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests package initializer to allow relative imports in tests."""
